requirements.txt
pandas
PyYAML
requests
dotenv
rich
mplfinance
yfinance
ib-insync
python-dotenv
pytz
six
